cambiando diseño de entidades
