+++ src/main/java/ar/com/plug/examen/domain/model/ClienteModel.java
@@ -4,23 +4,61 @@ import javax.persistence.*;
 
 @Entity
 @Table(name = "clientes")
-public class ClienteModel extends Persona{
+public class ClienteModel{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(unique = true, nullable = false)
-
+    private Long id;
+    private String nombre;
+    private String apellido;
+    private String telefono;
     private String ciudad;
     private String direccion;
     private Integer edad;
 
     public ClienteModel(Long id, String nombre, String apellido, String telefono, String ciudad, String direccion, Integer edad) {
-        super(id, nombre, apellido, telefono);
+        this.id = id;
+        this.nombre = nombre;
+        this.apellido = apellido;
+        this.telefono = telefono;
         this.ciudad = ciudad;
         this.direccion = direccion;
         this.edad = edad;
     }
 
     //getters y setters
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+    
+    public String getNombre() {
+        return nombre;
+    }
+
+    public void setNombre(String nombre) {
+        this.nombre = nombre;
+    }
+
+    public String getApellido() {
+        return apellido;
+    }
+
+    public void setApellido(String apellido) {
+        this.apellido = apellido;
+    }
+
+    public String getTelefono() {
+        return telefono;
+    }
+
+    public void setTelefono(String telefono) {
+        this.telefono = telefono;
+    }
+
     public String getCiudad() {
         return ciudad;
     }
