--- src/main/java/ar/com/plug/examen/domain/model/Persona.java
@@ -1,49 +1,0 @@
-package ar.com.plug.examen.domain.model;
-
-public class Persona {
-    //atributos
-    private Long id;
-    private String nombre;
-    private String apellido;
-    private String telefono;
-    //constructor
-    public Persona(Long id, String nombre, String apellido, String telefono) {
-        this.id = id;
-        this.nombre = nombre;
-        this.apellido = apellido;
-        this.telefono = telefono;
-    }
-    //getters y setters
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    public String getApellido() {
-        return apellido;
-    }
-
-    public void setApellido(String apellido) {
-        this.apellido = apellido;
-    }
-
-    public void setTelefono(String telefono){
-        this.telefono = telefono;
-    }
-
-    public String getTelefono(){
-        return telefono;
-    }
-    
-}
